+++ buession-springboot-pac4j/src/main/java/com/buession/springboot/pac4j/core/JWSAlgorithm.java
@@ -30,6 +30,7 @@ package com.buession.springboot.pac4j.core;
  * @author Yong.Teng
  * @since 2.0.1
  */
+@Deprecated
 public enum JWSAlgorithm {
 	HS256(com.nimbusds.jose.JWSAlgorithm.HS256.getName(), com.nimbusds.jose.JWSAlgorithm.HS256),