+++ buession-springboot-session/src/main/java/com/buession/springboot/session/autoconfigure/SessionConfiguration.java
@@ -35,6 +35,7 @@ import org.springframework.session.data.redis.config.annotation.web.http.EnableR
  */
 @Configuration(proxyBeanMethods = false)
 @ConditionalOnWebApplication
+@Deprecated
 public class SessionConfiguration {
 
 	@Configuration